Separate authentication logic to a new class
